+++ src/test/kotlin/com/nftco/flow/sdk/models/FlowPublicKeyTest.kt
@@ -1,0 +1,4 @@
+package com.nftco.flow.sdk.models
+
+class FlowPublicKeyTest {
+}
